Rename framework_test back to rpc
